Fixed upsell link on webinars page. Broke out the page into two parts. Removed featured box on job board. Reference #185, #184

+++ app/Http/Controllers/Auth/LoginController.php
@@ -45,11 +45,13 @@ class LoginController extends Controller
     {
         // TODO Check membership status
         $stripe_user = StripeServiceProvider::getCustomer($user);
-        if ($stripe_user->delinquent || $stripe_user->subscriptions->total_count < 1) {
-            // Remove clubhouse role from user
-            $role = RoleUser::where(array(array('role_code', 'clubhouse'), array('user_id', $user->id)))->first();
-            if ($role) {
-                $role->delete();
+        if (!is_null($stripe_user)) {
+            if ($stripe_user->delinquent || $stripe_user->subscriptions->total_count < 1) {
+                // Remove clubhouse role from user
+                $role = RoleUser::where(array(array('role_code', 'clubhouse'), array('user_id', $user->id)))->first();
+                if ($role) {
+                    $role->delete();
+                }
             }
         }
     }
